backend/requirements.txt
fastapi
uvicorn
websockets
openai
pandas
numpy
pydantic
python-dotenv
requests